Errors, Logging and Routes moved to proper folders
