+++ src/application/get_match_status_usecase.py
@@ -1,8 +1,8 @@
 from uuid import UUID
 
+from src.domain.logging.logger_interface import LoggerInterface
 from src.domain.models.status import Status
 from src.domain.repositories.match_database_repository import MatchDatabaseRepository
-from src.domain.services.logger_interface import LoggerInterface
 from src.domain.services.match_service import MatchService
