+++ src/application/make_movement_usecase.py
@@ -1,6 +1,6 @@
+from src.domain.logging.logger_interface import LoggerInterface
 from src.domain.models.movement import Movement
 from src.domain.repositories.match_database_repository import MatchDatabaseRepository
-from src.domain.services.logger_interface import LoggerInterface
 from src.domain.services.match_service import MatchService
