+++ src/domain/exception/errors.py
