+++ src/infra/repositories/in_memory_repository.py
@@ -1,8 +1,8 @@
 from uuid import UUID
 
+from src.domain.logging.logger_interface import LoggerInterface
 from src.domain.models.match import Match
 from src.domain.repositories.match_database_repository import MatchDatabaseRepository
-from src.domain.services.logger_interface import LoggerInterface
 
 
 class InMemoryRepository(MatchDatabaseRepository):
